리팩토링 -> FeedSectionModel 변경
- 이전 커밋과 동일하게 SectionDisplayModel 기반으로 변경하였습니다.
